% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_control_cnv.R
\name{add_control_cnv}
\alias{add_control_cnv}
\title{Add patient to Control DB}
\usage{
add_control_cnv(path_dir)
}
\arguments{
\item{path_dir}{path of the directory of the patient that will be analyzed. It must contain R1 and R2 (fasta/q files).}
}
\value{
Control DB updated
}
\description{
Saves the counts of the lectures of each gene in a control data base
which columns are the ids of the normal patients. It also changes the RDS file
where the databases are saved.
}
\examples{
ControlDB <- AddControl( id="123",  path_dir ="/home/daniela/Patients/123/123.fastq")
}
