% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_patient_cnv.R
\name{add_patient_cnv}
\alias{add_patient_cnv}
\title{Add patient to Patients DB}
\usage{
add_patient_cnv(path_dir)
}
\arguments{
\item{path_dir}{path of the directory of the patient that will be analyzed. It must contain R1 and R2 (fasta or fastq files).}

\item{id}{number or character that indicates which patient will be added.}
}
\value{
Patients DB updated
}
\description{
Saves the counts of the lectures of each gene in a data base
which columns are the ids of the patients. It also changes the RDS file
where the databases are saved.
}
\examples{
PatientsDB <-  AddPatient( id="123",  path_dir ="/home/daniela/Patients/123/123.fastq")
}
