% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addToRDS.R
\name{add_to_RDS}
\alias{add_to_RDS}
\title{Add to RDS}
\usage{
add_to_RDS(patient_XLSX_file)
}
\arguments{
\item{patient_XLSX_file}{es un ..}
}
\value{
frequency_of_interest frecuencias de
}
\description{
MitoR-.
}
