% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_XLSX.R
\name{analyzeVCF}
\alias{analyzeVCF}
\title{Modify the VCF file and delete the header}
\usage{
analyzeVCF(VCF_file)
}
\arguments{
\item{VCF_file}{Original VCF file}
}
\value{
List of variant information, and filter parameters (INDEL and SNP)
}
\description{
Discard the header of the VCF file and keep the variants information and the filter parameters (INDEL and SNP) used to generate the VCF
}
\examples{
homHet(list(0/0,0/1, 1/1))
}
