% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_from_fastq.R
\name{bam_from_fastq}
\alias{bam_from_fastq}
\title{Get BAM file from the FASTQ}
\usage{
bam_from_fastq(path_dir, keepFastQ = TRUE)
}
\arguments{
\item{path_dir}{path od the directory which name is the id of the patient and contains the FastQ files (R1 and R2).}

\item{keepFastQ}{indicates if the FastQ files are removed (if FALSE) or not (TRUE) after the conversion to BAM}
}
\value{
path of the BAM file
}
\description{
Generation of BAM file and it's index from the FastQ format.
}
\examples{
 
BAM_path <- bam_from_fastq(path_dir = "~/CNVAnalisisDefinitivo/PacientesMito/46608")
}
