% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_Indel_report.R
\name{bam_to_vcf}
\alias{bam_to_vcf}
\title{Creates a VCF file from a BAM file}
\usage{
bam_to_vcf(
  sorted_BAM,
  QD_SNPS = "< 2.0",
  FS_SNPS = "> 60.0",
  MQ_SNPS = "< 40.0",
  MQRankSum_SNPS = "< -12.5",
  ReadPosRankSum_SNPS = "< -8.0",
  QD_INDELS = "< 2.0",
  FS_INDELS = "> 200.0",
  ReadPosRankSum_INDELS = "< -20.0"
)
}
\value{
Returns the path of the created VCF file named "\if{html}{\out{<patient>}}_filtered_MitoR.vcf".
Several files are created and saved in a new directory called "MitoR" at the same directory as the R1 and R2 reads were located:
1 The VCF file
2 The VCF index file. It is saved
}
\description{
Selecting the path where the BAM file is saved, bam_to_vcf creates a VCF (Variant Call Format) file.
It uses a basic pipeline to create it, using the softwares: TrimGalore, BWA, PICARD, GATK, SamTools.
The default filters used for keeping or removing variants are:
For SNPs: 'QD < 2.0 || FS > 60.0 || MQ < 40.0 || MQRankSum < -12.5 || ReadPosRankSum < -8.0'
For indels: 'QD < 2.0 || FS > 200.0 || ReadPosRankSum < -20.0'
}
\examples{
bam_to_vcf("~/Github/MitochondriaAnalysis/MitoR/Patients/37019.bam")
}
