% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateXLSX (1).R
\name{buscarHMTVAR}
\alias{buscarHMTVAR}
\title{buscarHMTVAR}
\usage{
buscarHMTVAR(dataFrame)
}
\arguments{
\item{dataFrame}{It must have at least a REF, POS and ALT colomn name.}
}
\value{
lista
}
\description{
Takes information from the HMTVAR DataBase about the mutations we are interested in.
The taken data for each mutation is: clinvar, dbSNP, MitoMap, Omim, Disease, All_freq_h.
}
