% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa_kit.R
\name{check_gaps}
\alias{check_gaps}
\title{Check Presence of Gaps in a MSA}
\usage{
check_gaps(msa)
}
\arguments{
\item{msa}{A multiple sequence alignment object of mitochondrial CDS, a \code{DNAStringSet}
from the \code{Biostrings} package.}
}
\value{
Boolean. TRUE if gaps are presents, FALSE if not.
}
\description{
Check Presence of Gaps in a MSA
}
\examples{
\dontrun{
aln <- Biostrings::DNAStringSet(c( "ATGGCATCTACTTTGTATGACTATTGCAAGTGCCCATGGGTGACATCTGTAAGAAAGATGGGGATAAGCGCTGTAAGCTT", "ATGGCATCTACTTCGTATGACTATTGCAGAGTGCCCATG--------------GAAGACGGGGATAAGCGCTGTAAGCTT", "ATGGCATCTACTTTGTATGACTATTGCAGAGTGCCCATGGGTGACATCTGTAGAAAGATGGGGATAAGCGCTGTAAGCTT"))
# Check for gaps
check_gaps(aln)
}

}
