% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{checkDownloads}
\alias{checkDownloads}
\title{checkDownloads}
\usage{
checkDownloads()
}
\value{
When the download and installation is completed, a message will be displayed on the terminal
}
\description{
Every time the mitor package is load, this function will be called just to make sure you have all the needed softwares to make it run.
In case there is one o more of them missing, they will automatically be downloaded.
}
