% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateXLSX (2).R
\name{checkHMTVAR_online}
\alias{checkHMTVAR_online}
\title{Check if there is a MTVAR value to complete}
\usage{
checkHMTVAR_online()
}
\description{
When doing the analysis of a patient you must be online. Otherwise MitoR won't be able to look for the mutations information at HMTVAR.
In case you analysed it offline, as soon as you call the MitoR library again, MitoR will check if there are any Offline values and will try to look for them online.
}
