% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplogrep.R
\name{classify_haplogroup}
\alias{classify_haplogroup}
\title{Assign the Haplogroups of a set of human mtDNA Sequences}
\usage{
classify_haplogroup(sequences)
}
\arguments{
\item{sequences}{\code{DNAStringSet}. Set of human mitochondrial DNA sequences to classify.}
}
\value{
A table.
\describe{
\item{SampleID}{The ID of the sequence.}
\item{Haplogroup}{The predicted haplogroup.}
\item{Quality}{The quality score of the prediction.}
\item{Range}{The range of the sequence used for prediction.}
}

Output example:
\preformatted{
      SampleID   Haplogroup Rank Quality             Range
1   DQ856316.1        HV1b2    1  1.0000           1-16569
2   EU603401.1        K1a4a    1  0.9272           1-16569
3   KX681447.1          H7e    1  0.9610           1-16569
4   KX702230.1           H7    1  0.8705           1-16569
}
}
\description{
\code{classify_haplogroup} classify mitochondrial DNA haplogroups using \href{https://haplogrep.readthedocs.io/en/latest/installation/}{Haplogrep 3}
}
\details{
\code{classify_haplogroup} invokes \emph{haplogrep3}, which has to be previously installed on the user's machine.
For a step-by-step guide on installing and setting up haplogrep3, visit \url{github}.

The function might take 30-40 min for large datasets (~60000 sequences) and it runs in background.

For more information regarding the software and the results visit \url{https://haplogrep.readthedocs.io/en/latest/installation/}
}
