% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_from_counts.R
\name{cnv_from_counts}
\alias{cnv_from_counts}
\title{Get CNV calls from gene counts}
\usage{
cnv_from_counts(
  id,
  bed,
  ControlDB,
  PatientsDB,
  minoverlap = 1e-04,
  transit = 1
)
}
\arguments{
\item{id}{number or character that indicates which patient will be analyzed.
It has to be the same as its id in the data base.}

\item{bed}{bedfile that has information about the genes of the mitochondrial
genome. It's first 3 columns must be: chromosome, start, end. This file is
given on this package as bedfileMito.RDS in the Data folder.}

\item{ControlDB}{database in dataframe format that contains the counts of the
genes of normal patients that are used as a reference for making the analysis.}

\item{PatientsDB}{database in dataframe format that contains the counts of
the genes of any patient (normal or not) that has been analyzed.}

\item{minoverlap}{is a decimal number which indicates the needed percentage
of overlap from which it can be declare that a region has suffered a variation.}

\item{transit}{transition probability: Transition probability of the hidden Markov Chain
from the normal copy number state to either a deletion or a duplication.
The default (0.0001) expect approximately 20 CNVs genome-wide.}
}
\value{
CNV calls (dataframe that indicated the CNVs detected).
}
\description{
Using the ExomeDepth algorithm to detect deletions or duplications
of regions within the genes of the patient by comparing it against the normal
patients (my.ref.samples).
}
\examples{
CNV_calls_102 <- CNVfromCounts(102, bed_file, ControlDB, PatientsDB, minoverlap=0.001, transit = 0,7 )
}
