% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_Indel_report (1).R
\name{combine_workbooks}
\alias{combine_workbooks}
\title{combine_workbooks}
\usage{
combine_workbooks(workbooks, sheetNames)
}
\arguments{
\item{workbooks}{List of workbooks}
}
\value{
One single workbook containing all the initial ones as different sheets.
}
\description{
By giving a list of workbooks, combine_workbooks merges all of them into one single workbook. Each initial workbook is loaded as a single sheet in the resulting workbook.
}
