% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_patient_cnv.R
\name{delete_patient_cnv}
\alias{delete_patient_cnv}
\title{Remove a patient from Patients DB or Control DB.}
\usage{
delete_patient_cnv(id, n)
}
\arguments{
\item{id}{of the patient that will be removed.}

\item{n}{string that indicates  the database that will be changed. It must be
either "Control" or "Patients".}
}
\value{
DB updated
}
\description{
Removes the patient from the database indicated. Updates the RDS file where all databases are saved.
}
\examples{
PatientsDB <- RemoveAnyPatient( id= "37556", n = "Patients")
ControlDB <- RemoveAnyPatient( id= "37556", n = "control")
}
