% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_patient_mut.R
\name{delete_patient_mut}
\alias{delete_patient_mut}
\title{Remove a Patient's Analysis from the MitoR DataBase.}
\usage{
delete_patient_mut(numPaciente, keepXLSX = TRUE)
}
\arguments{
\item{numPaciente}{Patient's ID or title of the patient. In case you don't remember its name, we recommend you to look for it at the second page of the XLSX DataBase.
In case you still don't find it, run the function using a random number as numPaciente and the system will provide you with a list of the available patients.}

\item{keepXLSX}{After removing the patient from the DataBase, it will also remove the individual XLSX analysis created for the patient. TRUE by default.}
}
\value{
The patient's XLSX file containing the analysis will be removed. The patients mutations will also be removed from the DataBase if wanted.
}
\description{
After adding several patients analysis to the DataBase, you might want to remove some of them. This function allows you to remove
the patient from the DataBase. You also have the possibility to delete the complete analysis (except for the VCF file).
}
\examples{
deletePatient(37339)
deletePatient(37339, keepXLSX = FALSE)
}
