% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_RDS.R
\name{deleteFreq_MitoR}
\alias{deleteFreq_MitoR}
\title{Deletes a patient from the MitoR DB}
\usage{
deleteFreq_MitoR(patient_ID)
}
\arguments{
\item{patient_ID}{Patient's ID}
}
\value{
Updated MitoR DB with fixed values of frequencies and counts
}
\description{
Fixes the frequencies and count values by going through the patient's mutations element of the RDS list. Mutations with count = 0 are deleted.
}
