% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa_kit.R
\name{detect_disease_rna}
\alias{detect_disease_rna}
\title{Detect Reported and Confirmed Disease Associated Mutation}
\usage{
detect_disease_rna(msa, target)
}
\arguments{
\item{msa}{A multiple sequence alignment object, a \code{DNAStringSet}
from the \code{Biostrings} package.}

\item{target}{Character. Name of target sequence.}
}
\value{
Data frame with detected mutations. \emph{NULL} if none
}
\description{
This function detects if a target sequence presents confirmed or reported mutation associated with mitochondrial diseases in rRNAs and tRNAs.
The mutations that are checked can be found in the \href{https://www.mitomap.org/foswiki/bin/view/MITOMAP/MutationsRNA}{MITOMAP} database.
}
