% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_cnv.R
\name{detection_cnv}
\alias{detection_cnv}
\title{Detect copy number variations (CNVs) of a patient sample.}
\usage{
detection_cnv(path_dir, minoverlap, transit)
}
\arguments{
\item{path_dir}{path of the directory of the patient that will be analyzed. It must contain R1 and R2 (fasta or fastq files).}

\item{minoverlap}{is a decimal number which indicates the needed percentage
of overlap from which it can be declare that a region has suffered a variation.}

\item{transit}{transition probability: Transition probability of the hidden Markov Chain from the normal copy number state to either a deletion or a duplication.
This parameter is taken from ExomeDepth algorithms.}
}
\value{
a list:
\itemize{
\item CNV calls: dataframe with information about CNVs.
\item Graph: visual representation of CNVs. Duplications colored in green and deletions colored in red.
}
}
\description{
Generates a dataframe and a visual representation of the CNVs detected.
}
