% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{disease_cds_df}
\alias{disease_cds_df}
\title{Reported and Confirmed Mitochondrial Associated Disease Mutations}
\format{
An object of class \code{data.frame} with 536 rows and 12 columns.
}
\usage{
disease_cds_df
}
\description{
This data frame contains information about reported and confirmed mitochondrial associated disease mutations in control and coding regions
as reported by \href{https://www.mitomap.org/foswiki/bin/view/MITOMAP/MutationsCodingControl}{MITOMAP}.
}
\details{
Visit \href{https://www.mitomap.org/foswiki/bin/view/MITOMAP/MutationsCodingControl}{MITOMAP} for more and clear details.
}
\keyword{datasets}
