% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{downloadHG38}
\alias{downloadHG38}
\title{downloadHG38}
\usage{
downloadHG38(mitor_sof)
}
\value{
The path where the .fasta reference file is located
}
\description{
Downloads and decompresses the HG38 mitochodnrial DNA reference.
}
