% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{downloadMitoRSoftwares}
\alias{downloadMitoRSoftwares}
\title{downloadMitoRSoftwares}
\usage{
downloadMitoRSoftwares(mitor_sof)
}
\value{
After each software is correctly downloaded and installed, a message from each one will be displayed on the terminal.
In case it occurs an error because of a missing package on your Linux OS, a message will be displayed on the terminal letting you know which packages must be available to download the software.
}
\description{
Downloads all the necessary softwares and the HG38 file reference to use the MitoR software.
}
