% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{downloadPICARD}
\alias{downloadPICARD}
\title{downloadPICARD}
\usage{
downloadPICARD(mitor_sof)
}
\value{
The path where the .exe file is located
}
\description{
Downloads and decompresses the PICARD software
}
