% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa_kit.R
\name{extract_trna}
\alias{extract_trna}
\title{Extract tRNA Genes from a Mitochondrial DNA Multiple Sequence Alignment}
\usage{
extract_trna(msa, coord_df)
}
\arguments{
\item{msa}{A multiple sequence alignment object, a \code{DNAStringSet}
from the \code{Biostrings} package.}

\item{coord_df}{Data frame. A data frame created using the function \code{genes_coord}.}
}
\value{
A \code{DNAStringSetList} object containing the extracted tRNA gene sequences,
with gene identifiers as names.
}
\description{
This function extracts transfer RNA (tRNA) genes from a multiple sequence alignment (MSA)
based on the coordinates provided in a data frame. The extracted tRNA sequences are returned
as a named \code{DNAStringSetList}.
}
