% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_Indel_report (1).R, R/SNP_Indel_report.R,
%   R/fasta_to_bam.R, R/generateVCF (1).R
\name{fasta_to_bam}
\alias{fasta_to_bam}
\title{fasta_to_bam}
\usage{
fasta_to_bam(path_dir)

fasta_to_bam(path_dir)

fasta_to_bam(path_dir)

fasta_to_bam(path_dir)
}
\arguments{
\item{path_dir}{path of the directory of the patient that will be analyzed.}

\item{path}{es el directorio donde tiene que buscar los dos archivos}
}
\value{
Returns the path of the created BAM file named "\if{html}{\out{<patient>}}_sortedR1R2.bam".

Nos devuelve tres elementos:
\itemize{
\item Archivo .fasta o .fastq R1
\item Archivo .fasta o .fastq R2
\item Nombre del paciente. Es decir, la identificacion que poseen los archivos R1 y R2 en comun
}

path of the sorted BAM file.

Nos devuelve tres elementos:
\itemize{
\item Archivo .fasta o .fastq R1
\item Archivo .fasta o .fastq R2
\item Nombre del paciente. Es decir, la identificacion que poseen los archivos R1 y R2 en comun
}
}
\description{
Selecting the path where the fasta or fastq files are saved, fasta_to_bam creates a BAM (Binary Alignment Map) file.

Al darle un path de un directorio, va a buscar los archivos R1 y R2 que ademas sean .fasta o .fastq

Given the path of a directory, this function will identify the R1 and R2 files
and will generate a SAM file, then a BAM and a sorted BAM file by applying commands from BWA,
Picard, Samtools and GATK and using the mitochondrial human genome reference.
All this softwares and reference are downloaded automathically with the installation of MitoR.

Al darle un path de un directorio, va a buscar los archivos R1 y R2 que ademas sean .fasta o .fastq
}
