% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_mtdna.R
\name{fetch_seq}
\alias{fetch_seq}
\title{fetch_seq}
\usage{
fetch_seq(query, dir_path, n = NULL)
}
\arguments{
\item{query}{NCBI query to download}

\item{dir_path}{path where to save the files}

\item{n}{number of sequences to be download}
}
\value{
nothing
}
\description{
fetch_seq
}
\examples{
ciao
}
