% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa_kit.R
\name{find_var_pos}
\alias{find_var_pos}
\title{Find Non Conserved Columns in a Multiple Sequence Alignment (MSA)}
\usage{
find_var_pos(msa, type = "AA")
}
\arguments{
\item{msa}{A multiple sequence alignment object, a \code{DNAStringSet}
from the \code{Biostrings} package.}

\item{type}{String. Type of MSA, amino acids ("AA") or nucleotides ("DNA").}
}
\value{
Matrix. The function returns a matrix that looks like this. The names of the columns coresponds to the
position in the MSA, while the rownames to the name of the sequence.

\if{html}{\out{<div class="sourceCode">}}\preformatted{ 14  29  30  31  32  33  36  37  38  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54  59
}\if{html}{\out{</div>}}

seq1 "T" "A" "G" "T" "G" "C" "A" "T" "G" "G" "T" "G" "A" "C" "A" "T" "C" "T" "G" "T" "A" "A" "G" "A" "T"
seq2 "C" "G" "A" "G" "T" "G" "C" "A" "T" "-" "-" "-" "-" "-" "-" "-" "-" "-" "-" "-" "-" "-" "-" "G" "C"
seq3 "T" "G" "A" "G" "T" "G" "C" "A" "T" "G" "G" "T" "G" "A" "C" "A" "T" "C" "T" "G" "T" "A" "G" "A" "T"
}
\description{
This function finds and extracts the non conserved columns in a MSA.
}
\examples{
\dontrun{
aln <- Biostrings::DNAStringSet(c( "ATGGCATCTACTTTGTATGACTATTGCAAGTGCCCATGGGTGACATCTGTAAGAAAGATGGGGATAAGCGCTGTAAGCTT", "ATGGCATCTACTTCGTATGACTATTGCAGAGTGCCCATG--------------GAAGACGGGGATAAGCGCTGTAAGCTT", "ATGGCATCTACTTTGTATGACTATTGCAGAGTGCCCATGGGTGACATCTGTAGAAAGATGGGGATAAGCGCTGTAAGCTT"))
names(aln) <- paste0("seq", 1:3)
find_var_pos(aln, type="DNA")
}
}
