% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateVCF.R
\name{fix_filter_values}
\alias{fix_filter_values}
\title{Fix the filter values}
\usage{
fix_filter_values(
  QD_SNPS,
  FS_SNPS,
  MQ_SNPS,
  MQRankSum_SNPS,
  ReadPosRankSum_SNPS,
  QD_INDELS,
  FS_INDELS,
  ReadPosRankSum_INDELS
)
}
\arguments{
\item{muchos}{blabla}
}
\value{
parametros
}
\description{
blabla
}
