% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cnv_report.R
\name{generate_cnv_report}
\alias{generate_cnv_report}
\title{Generation of report}
\usage{
generate_cnv_report(CNV_calls, graph, path_dir, transit)
}
\arguments{
\item{CNV_calls}{dataframe generated using AnalyzePatient function.}

\item{graph}{ggplot that represents the CNVs.}

\item{path_dir}{path of the directory of the patient that has been analized. It must contain R1 and R2.}

\item{transit}{transition probability: Transition probability of the hidden Markov Chain
from the normal copy number state to either a deletion or a duplication. Is the value used
in the generation of the CNV_calls indicated in the first parameter.}
}
\value{
path where the report has been saved.
}
\description{
Generates the excel with the results of the patient's analysis.
You will find each CNV detected with it's information such as size, start, end,
association with Franklin and Varsome, etc.
}
