% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_links.R
\name{generate_links}
\alias{generate_links}
\title{Generation of Links}
\usage{
generate_links(CNV_calls)
}
\arguments{
\item{CNV_calls}{dataframe created with ExomeDepth algorithm.}
}
\value{
CNV_calls with 2 new columns "Franklin" and "Varsome" that contain
the links that can be pasted in the browser to be redirected to any of
these databases and get all the information about t¡an especific variation
within the CNVs found in this case.
}
\description{
Generates links to databases for each variation
}
