% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateVCF (1).R
\name{generate_VCF}
\alias{generate_VCF}
\title{Creates a VCF file from reads R1 & R2 (fasta or fastq files)}
\usage{
generate_VCF(
  path,
  QD_SNPS = "< 2.0",
  FS_SNPS = "> 60.0",
  MQ_SNPS = "< 40.0",
  MQRankSum_SNPS = "< -12.5",
  ReadPosRankSum_SNPS = "< -8.0",
  QD_INDELS = "< 2.0",
  FS_INDELS = "> 200.0",
  ReadPosRankSum_INDELS = "< -20.0"
)
}
\arguments{
\item{keep_BAM=FALSE}{While creating the VCF file, a BAM file is created. If you want to keep it as a part of the analysis, it will be saved in the same directory as the VCF file.}
}
\value{
Returns a the path of the created VCF file named "\if{html}{\out{<patient>}}_filtered_MitoR.vcf".
Several files are created and saved in a new directory called "MitoR" at the same directory as the R1 and R2 reads were located:
1 The VCF file
2 The VCF index file. It is saved
3 The BAM file (Optional. Default = FALSE)
In case you want to keep the BAM file, it will be placed at a new directory named of "BAMFile" at the same location as the VCF file.
}
\description{
Selecting the directory where the fasta or fastq files are saved, generate_VCF creates a VCF (Variant Call Format) file.
It uses a basic pipeline to create it, using the softwares: TrimGalore, BWA, PICARD, GATK, SamTools.
The filters used for keeping or removing mutation data are:
For SNPs: 'QD < 2.0 || FS > 60.0 || MQ < 40.0 || MQRankSum < -12.5 || ReadPosRankSum < -8.0'
For indels: 'QD < 2.0 || FS > 200.0 || ReadPosRankSum < -20.0'
}
\examples{
R1_R2_directory_path = "~/Github/MitochondriaAnalysis/MitoR/Patients/37019"
generate_VCF(R1_R2_directory_path);
generate_VCF(R1_R2_directory_path, keep_BAM = TRUE)
}
