% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_XLSX.R
\name{generate_XLSX}
\alias{generate_XLSX}
\title{Convert VCF file to XLSX file}
\usage{
generate_XLSX(VCF_file)
}
\arguments{
\item{addToDataBase=TRUE}{While creating the XLSX file, the information is added to a DataBase.
If you want to keep add it as part of the analysis, it will be saved in the upper directory of the R1 and R2 files of the analyzed patient.}
}
\value{
XLSX file
}
\description{
It generates a XLSX file with the information of the mutations available at the VCF file.
Creates a XLSX file from a VCF file. If you do not have one, we recommend you to see mitor::generateVCF()
By doing this, the genetist or the bioinformatitian is able to have the information on a more common file type.
Information about the mutation is also taken from HMTVAR. To use this function you must be online.

The function has also the possibility to store the data in a database automatically generated. The database will also be a XLSX file.
}
\examples{
generate_XLSX("../MitoR/patient1.vcf", addToDataBase=FALSE)
generate_XLSX("../MitoR/patient1.vcf")
}
