% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_gene.R
\name{generateTXT_gene}
\alias{generateTXT_gene}
\title{Generation of TXT}
\usage{
generateTXT_gene(numPaciente)
}
\arguments{
\item{numPaciente}{blablaba}
}
\value{
path_for_positions, mutations_at_genes
}
\description{
MitoR-.
}
