% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_gene.R
\name{genes_of_mutations}
\alias{genes_of_mutations}
\title{Modifies the patients mutation table.}
\usage{
genes_of_mutations(mutations, gene_start_end)
}
\arguments{
\item{mutations}{DataFrame of colomns REF-POS-ALT}

\item{gene_start_end}{BED file with modified length depending on the insertions/deletions of the patient}
}
\value{
Updated BED table as a DataFrame object. Containing colomns: Interval, gene name, mutations found (REF-POS-ALT)
}
\description{
Makes modifications to the mutation's table
}
