% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3d_analysis.R
\name{get_coords}
\alias{get_coords}
\title{Get 3D Coordinates of Protein Variants}
\usage{
get_coords(prot_list, t = NULL)
}
\arguments{
\item{prot_list}{List. Protein list generated by translating the output of \emph{extract_cds}.}

\item{t}{Double. Frequency threshold, select positions in the MSA which its most conserved amino acid/nucleotide has a frequency of at most \emph{t}.}
}
\value{
Data frame containing the 3D coordinates of the positions of non-synonymous variants
}
\description{
Get 3D Coordinates of Protein Variants
}
