% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBAMBAI.R
\name{getBAMBAI}
\alias{getBAMBAI}
\title{Get the BAM file from the patient selected by the path.}
\usage{
getBAMBAI(path_dir)
}
\arguments{
\item{path_dir}{path of the directory of the patient that will be analyzed. It must contain R1 and R2 (fasta or fastq files).}
}
\value{
bam.control path of the sorted BAM file
}
\description{
Check if there are BAM or BAI files already generated on the directory. If there are not, generates them.
}
