% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gnomAD_df}
\alias{gnomAD_df}
\title{gnomAD v3.1 Mitochondrial DNA Variants}
\format{
A \code{data frame} of dimenison 10850X99:
\describe{
\item{\code{CHROM}}{Character. The chromsome, "chrM" for all the entries.}
\item{\code{POS}}{Integer. The position of the variant in the rCRS.}
\item{\code{ID}}{Character. refSNP id for the variant. Not always available.}
\item{\code{AF_hom}}{Double. Frequency of the variant in homoplasmic state in the database. }
\item{\code{Allele}}{Character. Allele state of the alternative variant. }
\item{\code{Consequence}}{Character. Type of variant: "intergenic_variant", "non_coding_transcript_exon_variant", "start_lost",
"stop_gained&start_lost",     "synonymous_variant" ,
"missense_variant",
"frameshift_variant",
"stop_gained",
"frameshift_variant&stop_lost" ,
"stop_lost",
"stop_retained_variant",
"coding_sequence_variant&3_prime_UTR_variant",
"inframe_deletion",
"incomplete_terminal_codon_variant&coding_sequence_variant". }
\item{\code{Protein_position}}{Numeric. Position of the amino acid in the protein.}
\item{\code{AA_ref}}{Character. Amino acid (one letter code) in the rCRS.}
\item{\code{AA_alt}}{Character. Amino acid (one letter code) in the variant.}
}
}
\source{
\href{https://gnomad.broadinstitute.org/news/2020-11-gnomad-v3-1-mitochondrial-dna-variants/}{gnomAD v3.1 Mitochondrial DNA Variants}
}
\usage{
gnomAD_df
}
\description{
This dataframe contains the Mitochondrial DNA (mtDNA) variants for \href{https://gnomad.broadinstitute.org/news/2020-11-gnomad-v3-1-mitochondrial-dna-variants/}{gnomaAD}.
}
\details{
This data frame contains information of the human mitochondrial variants as stored in the \href{https://gnomad.broadinstitute.org/news/2020-11-gnomad-v3-1-mitochondrial-dna-variants/}{gnomaAD}.
Since the file on the database is a vcf file, this data frame allows the user to navigate through them in a more
user friendly way. Moreover, the data frame presents only the variants that passed the quality filter pipeline, as described in their \href{https://gnomad.broadinstitute.org/news/2020-11-gnomad-v3-1-mitochondrial-dna-variants/}{website}.
}
\keyword{datasets}
