% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_XLSX.R
\name{homHet}
\alias{homHet}
\title{homHet}
\usage{
homHet(GT)
}
\arguments{
\item{GT}{List of GATK parameters as 0/0, 0/1, 1/1}
}
\value{
The same list (if it is coming from the generateXLSX function, it is a dataFrame colomn) with the replaced meaning.
}
\description{
Given the GATK parameters as 0/0, 0/1, 1/1, this function writes its meaning regarding the type of mutation.
}
\examples{
homHet(list(0/0,0/1, 1/1))
}
