% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_gene.R
\name{indels_in_gene}
\alias{indels_in_gene}
\title{Modifies the BED table. Makes a gene longer or shorter depending on the patients variants.}
\usage{
indels_in_gene(mutations, start_end_gene)
}
\arguments{
\item{mutations}{DataFrame of colomns REF-POS-ALT}

\item{start_end_gene}{BED file - sprintf("\%s/MitoRSoftware/bedfileMito.RDS", mitor_files)}
}
\value{
Updated BED table as a DataFrame object
}
\description{
Depending on the position of the insertion or deletion of the variant, it changes the length of the gene's start and end
}
