% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_db.R
\name{initialize_db}
\alias{initialize_db}
\title{Initialize Data Bases}
\usage{
initialize_db()
}
\value{
list with 3 databases
}
\description{
Generation of 3 empty databases: Control(normal references),
Patients (analyzed) and Genes (genes detected with variation) if it is the first
time that the user uses this package. Else, it reads the databases already
created.
}
\examples{
initial <- InitializeDB()
ControlDB <- initial[[1]]
PatientsDB <- initial[[2]]
GenesDB <- initial[[3]]
}
