% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3d_analysis.R
\name{map_variants}
\alias{map_variants}
\title{Map Non-Synonymous Variants To Protein Structure}
\usage{
map_variants(variants, complex = NULL)
}
\arguments{
\item{variants}{Data frame. Data frame containing the location of the variants, generated by the function \emph{get_coords}}

\item{complex}{String. Roman number corresponding to the electron transport chain complexes of "I", "III", "IV", "V"}
}
\description{
\code{map_variants()} maps the location of the variants on the 3D structure of the complexes using PyMol.
The variants location are highlighted in red.
}
