% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{my_msa}
\alias{my_msa}
\title{Multiple Sequence Alignment of Human mtDNA}
\format{
An object of class \code{DNAStringSet} of length 301.
}
\usage{
my_msa
}
\description{
This \emph{DNAStringSet} corresponds to a multiple sequence alignment of the rCRS and the first 300 mtDNA sequences corresponding to the query:
}
\details{
"(016500\link{SLEN}:016600\link{SLEN}) AND mitochondrion\link{FILT} AND txid9606\link{orgn} NOT txid63221\link{orgn} NOT txid741158\link{orgn} NOT NC_012920.1\link{accn}"

and fetched with the \emph{fetch_mtdna} function. It has been aligned using Clustal Omega.

It can be used as an example to see how the package works.
}
\keyword{datasets}
