% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_gene.R
\name{pileupPlot}
\alias{pileupPlot}
\title{Pile up Plot}
\usage{
pileupPlot(numPaciente, plot_mutation = TRUE, range = 5)
}
\arguments{
\item{numPaciente, }{plot_mutation = TRUE, range = 5}
}
\description{
MitoR-.
}
