% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cnv.R
\name{plot_cnv}
\alias{plot_cnv}
\title{Generation of the graph of the report}
\usage{
plot_cnv(CNV_calls, bed)
}
\arguments{
\item{CNV_calls}{dataframe generated using AnalyzePatient function. It has the information about the CNVs detected.}

\item{bed}{is the bedfile that is loaded in this library. You can check it out by coding data("bedfileMito"). It contains information about the 37 genes in the mitochondrial genome}
}
\value{
graph
}
\description{
Generates a graph that represents the CNVs detected
}
