% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_gene.R
\name{plot_gene_design}
\alias{plot_gene_design}
\title{Designs the gene pileup plot}
\usage{
plot_gene_design(allelic_coverage, mutations_at_genes)
}
\arguments{
\item{allelic_coverage}{DataFrame of colomns POS-REFCounts-ALT_counts-REF-ALT-TotalCounts}

\item{mutations_at_genes}{DataFrame of colomns: Interval, gene name, mutations found (REF-POS-ALT)}
}
\value{
Recorded gene pileup plot
}
\description{
Checking the allelic coverage for the gene and the mutations belonging to the gene, this function designs the plot, its lines, text and legends
}
