% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_gene.R
\name{plot_mutation_design}
\alias{plot_mutation_design}
\title{Designs the mutation pileup plot}
\usage{
plot_mutation_design(data_mutation_plot, alternative_base, range)
}
\arguments{
\item{data_mutation_plot}{DataFrame of colomns POS-REFCounts-ALT_counts-REF-ALT-TotalCounts for the mutated position and the \if{html}{\out{<range>}} closests nucleotides}

\item{alternative_base}{ALT of the mutation}
}
\value{
Recorded mutation pileup plot
}
\description{
Checking the allelic coverage for the mutated position and the \if{html}{\out{<range>}} closests nucleotides, his function designs the plot, its lines, text and legends
}
