% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{position_entropy_prot}
\alias{position_entropy_prot}
\title{Find Shannon Entropy for each position of a MSA of Proteins}
\usage{
position_entropy_prot(msa)
}
\arguments{
\item{msa}{A multiple sequence alignment object, a \code{AAStringSet}
from the \code{Biostrings} package.}
}
\value{
A vector with the entropy value for each position of the MSA.
}
\description{
Find Shannon Entropy for each position of a MSA of Proteins
}
