% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predict_af3}
\alias{predict_af3}
\title{Predicting Protein Structure With AlphaFold3 Server}
\usage{
predict_af3(seq = NULL, name = NULL, dir = NULL)
}
\arguments{
\item{seq}{String. Amino acids sequence of the protein, only canonical AAs are accepted. AAString object are accepted.}

\item{name}{String. Name of the zip file containing the prediction to be created.}

\item{dir}{String. Path of the directory where to save the zip file with the prediction. If it doesn't already exist, it will be created.}
}
\value{
The function
}
\description{
Predicting Protein Structure With AlphaFold3 Server
}
