% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rCRS}
\alias{rCRS}
\title{Cambridge Reference Sequence - Human Mitochondrial Genome}
\format{
A \code{DNAStringSet} object of length 1:
\describe{
\item{\code{width}}{Integer. The total length of the DNA sequence, 16,569 base pairs.}
\item{\code{seq}}{Character. The complete nucleotide sequence of the human mitochondrial genome.}
\item{\code{names}}{Character. The accession number and identifier of the sequence, \code{"NC_012920"}.}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/nuccore/NC_012920.1/}
}
\usage{
rCRS
}
\description{
This object contains the full DNA sequence of the revised Cambridge Reference Sequence (rCRS), representing the human mitochondrial genome.
}
\details{
It serves as the reference for sequence alignment, annotation, and comparative studies in mitochondrial genetics.
For more details check this \href{https://www.mitomap.org/foswiki/bin/view/MITOMAP/HumanMitoSeq}{website}
}
\keyword{datasets}
