% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa_kit.R
\name{save_genes}
\alias{save_genes}
\title{Save Genes from a mtDNA Multiple Sequence Alignment in FASTA files}
\usage{
save_genes(msa, coord_df, dir = "data")
}
\arguments{
\item{msa}{A multiple sequence alignment object, a \code{DNAStringSet}
from the \code{Biostrings} package.}

\item{coord_df}{Data frame. A data frame created using the function \code{genes_coord}.}

\item{dir}{String. Name of the directory where to save the extracted genes.}
}
\description{
This function extracts the 37 mitochondrial genes (protein-coding genes, tRNAs, and rRNAs)
from a multiple sequence alignment (MSA) and saves them as FASTA files organized by type.
Each FASTA file will correspond to the subset of the MSA containing a specific gene.
}
