% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitomap.R
\name{search_allele}
\alias{search_allele}
\title{Search Positions Information in the Mitomap Database}
\usage{
search_allele(pos = NULL)
}
\arguments{
\item{pos}{Vector. Numeric vector of mitochondrial positions (max. 100).}
}
\value{
A list of data frames. Each data frame corresponds to the variants in Coding Region Sequence Variants,
Control Region Sequence Variants, Reported Mitochondrial DNA Base Substitution Diseases.
}
\description{
search_allele allows the user to programmatically retrieve information regarding human mitochondrial varaints at a given position as shown in the
MITOMAP database. It is equivalent to manually searching it in \href{https://www.mitomap.org/allelesearch.html}{webserver},
but with the advantage of returning results in a structured data frame format, making it easier to analyze.
}
\examples{
\dontrun{
  pos <- 13800:13900
 example <- search_allele(pos)
 test
  }
}
