% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_XLSX.R
\name{searchHMTVAR}
\alias{searchHMTVAR}
\title{buscarHMTVAR}
\usage{
searchHMTVAR(dataFrame)
}
\arguments{
\item{dataFrame}{It must have at least a REF, POS and ALT colomn name.}
}
\value{
list with: clinvar, dbSNP, MitoMap, Omim, Disease, All_freq_h
}
\description{
Takes information from the HMTVAR DataBase about the mutations we are interested in.
The taken data for each mutation is: clinvar, dbSNP, MitoMap, Omim, Disease, All_freq_h.
In case you are offline, the report will show an "Offline" value in the cells related to HmtVar
In case you the HmtVar website is down, the report will show an "HmtVar Down" value in the cells related to HmtVar
}
