% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateVCF.R
\name{searchPatient}
\alias{searchPatient}
\title{searchPatient}
\usage{
searchPatient(path)
}
\arguments{
\item{path}{es el directorio donde tiene que buscar los dos archivos}
}
\value{
Nos devuelve tres elementos:
\itemize{
\item Archivo .fasta o .fastq R1
\item Archivo .fasta o .fastq R2
\item Nombre del paciente. Es decir, la identificacion que poseen los archivos R1 y R2 en comun
}
}
\description{
Al darle un path de un directorio, va a buscar los archivos R1 y R2 que ademas sean .fasta o .fastq
}
\examples{
searchPatient(~/MitoR/Pacientes/39337)
}
