% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_Indel_report.R
\name{SNP_Indel_Analyze}
\alias{SNP_Indel_Analyze}
\title{Analyze a DNA sequence to look for SNPs & INDELs}
\usage{
SNP_Indel_Analyze(
  path,
  QD_SNPS = "< 2.0",
  FS_SNPS = "> 60.0",
  MQ_SNPS = "< 40.0",
  MQRankSum_SNPS = "< -12.5",
  ReadPosRankSum_SNPS = "< -8.0",
  QD_INDELS = "< 2.0",
  FS_INDELS = "> 200.0",
  ReadPosRankSum_INDELS = "< -20.0"
)
}
\arguments{
\item{path}{Either the BAM file or the directory containing the fasta or fastq files}

\item{...}{Filters applied to the analysis of SNPs or INDELs to keep or remove a detected variant. Default numbers are the recommended by GATK}
}
\value{
XLSX file containing several sheets are returned:
\itemize{
\item SNP and INDEL report: for EACH variant you will get hyperlinks to Franklin, Varsome and dbSNP databases, information from HMTVAR database, filtering information from GATK and PICARD, allele depth and coverage information.
\item Plot report: for EACH mutated gene you will get a single plot showing the read coverage. And for each mutation you will also get a single plot with the same information, showing the sequence with the centered variant and the next and previous nucleotides (range of 5 nucleotides)
\item Softwares report: Information regarding the versions, dates of analysis and data from the used softwares (BWA, GATK, PICARD, SamTools)
}
}
\description{
It generates a XLSX file with the SNPs & INDELs found, with information about them from HMTVAR. It also creates hyperlinks to Franklin, Varsome, dbSNP for each variant.
It will also generate a VCF file which will be saved in the same directory as the input path is set. The VCF is filtered with the setted parameters.
The analysis will be automatically loaded in MitoR DataBase.
}
\examples{
SNP_Indel_report("../MitoR/patient1.vcf")
generate_XLSX("../MitoR", QD_SNPS = < 2.5, FS_SNPS = > 45.0)
}
