% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa_kit.R
\name{translation}
\alias{translation}
\title{Translate MSA Corresponding to a CDS}
\usage{
translation(msa)
}
\arguments{
\item{msa}{A multiple sequence alignment object of mitochondrial CDS, a \code{DNAStringSet}
from the \code{Biostrings} package. Typically a DNAStringSet object contained in the list
generated by the function \code{extract_cds}.}
}
\value{
A \code{DNAStringSet} from the \code{Biostrings} package. The result will not necessarily be a MSA, since gaps are removed in the translation.
}
\description{
This function translates a given MSA of mitochondrial CDS into the corresponding AA sequences.
}
\note{
CDS MSA containing gaps are not reccomended for this function.
}
