% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_DB_cnv.R, R/update_genesdb.R
\name{update_DB_cnv}
\alias{update_DB_cnv}
\title{Add another patient's information to Exons DB}
\usage{
update_DB_cnv(CNV_calls, id, ControlDB, PatientsDB, GenesDB)

update_DB_cnv(CNV_calls, id, ControlDB, PatientsDB, GenesDB)
}
\arguments{
\item{CNV_calls}{dataframe which contains the information about the CNVs of the new patient}

\item{id}{number or character that indicates which patient is.}

\item{e}{name of the RDS file  for GenesDB}
}
\value{
Genes DB updated

Genes DB updated
}
\description{
Adds the information about the CNVs of exons of a new patient
and keeps track of the frequency and type of variation of each exon.

Adds the information about the CNVs of exons of a new patient
and keeps track of the frequency and type of variation of each exon.
}
