% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateMitoRStats_HMTVAR.R
\name{update_freq_MitoR}
\alias{update_freq_MitoR}
\title{Adds or updates MitoR DB frequencies to the patient report}
\usage{
update_freq_MitoR(numPaciente)
}
\arguments{
\item{numPaciente}{Patient ID.}
}
\value{
Updated report and a message telling the path it is saved at.
}
\description{
Takes the MitoR DB frequencies from all the previous analysis and loads them into the patient report.
It is important that you already have the analysis previously made. Otherwise, the function will analyze the patient from the VCF.
If you did not keep the XLSX report nor the VCF file, you must do the analysis again with the patient's fasta/fastq files or its BAM file.
}
\examples{
update_freq_MitoR("47286")
}
