% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateMitoRStats_HMTVAR (1).R
\name{updateFreqMutMitoR}
\alias{updateFreqMutMitoR}
\title{Update the MitoR-DataBase frequency}
\usage{
updateFreqMutMitoR(numPaciente)
}
\arguments{
\item{numPaciente}{blabla}
}
\description{
MitoR-
}
