% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateMitoRStats_HMTVAR.R
\name{updateMitoRStats_HMTVAR}
\alias{updateMitoRStats_HMTVAR}
\title{Update the MitoR-DataBase stats taken from HMTVAR}
\usage{
updateMitoRStats_HMTVAR(AddToPatients = FALSE)
}
\arguments{
\item{AddToPatients}{In case you want to update the information of the individual XLSX files from the patients as well, keep AddToPatients = TRUE.
Please be aware that this will take a while if you have many patients to update.
In case you want to update only one patient or just some of them, we recommend you to create the XLSX file from the VCF again. Check generateXLSX() for more information.}
}
\value{
The DataBase XLSX file will automatically be updated. The same will happen with the patients XLSX files in case of AddToPatients = TRUE
}
\description{
MitoR-DataBase takes information for each mutation from the HMTVAR DataBase. This DataBase is constantly updated.
We recommend to update the information of MitoR-DataBase at least once a month. With this function, the MitoR DataBase will be updated with the last
available information about each mutation from HMTVAR.
}
\examples{
updateMitoRStats_HMTVAR()
updateMitoRStats_HMTVAR(addToPatients = TRUE)
}
