% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateMitoRStats_HMTVAR.R
\name{view_patient}
\alias{view_patient}
\title{View Patient's Report}
\usage{
view_patient()
}
\value{
message that indicates that the report has been updated and downloaded.
}
\description{
Downloads the xlsx report of a specific patient with the data updated.
}
\examples{
view_patient()
}
