% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_XLSX.R
\name{xlsx_report_design}
\alias{xlsx_report_design}
\title{Design the MitoR report workbook}
\usage{
xlsx_report_design(filtered_DF, filtered_DF2)
}
\arguments{
\item{filtered_DF}{DataFrame containing the SNP & INDEL report}

\item{filtered_DF2}{DataFrame containing the software reports}
}
\description{
Designs the WB of SNP, INDEL and software reports. It creates the hyperlinks to Franklin, Varsome and dbSNP DB
}
