% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_patient_cnv.R
\name{add_patient_cnv}
\alias{add_patient_cnv}
\title{Add patient to Patients DB}
\usage{
add_patient_cnv(path_dir)
}
\arguments{
\item{path_dir}{path of the directory of the patient that will be analyzed. It must contain R1 and R2 (fasta or fastq files).}
}
\value{
Patients DB updated
}
\description{
Saves reads' counts of each gene from the sequence of the patient in a dataset.
}
\examples{
PatientsDB <-  AddPatient( path_dir ="/home/sam/Patients/123/123.fastq")
}
