% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_RDS.R
\name{add_to_RDS}
\alias{add_to_RDS}
\title{Add the patient's analysis to the MitoR DB}
\usage{
add_to_RDS(filtered_DF, filtered_DF2, patient_ID)
}
\arguments{
\item{filtered_DF}{Patient's SNP & INDEL report}

\item{filtered_DF2}{Patient's software report}

\item{patient_ID}{Patient's ID}
}
\value{
Vector of frequencies from MitoR DB to add to the patient's report
}
\description{
Updates your MitoR DB with the new patient's analysis. Fixes the frequencies values, adds the new mutations and saves the date of the last MitoR DB update.
}
