% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_RDS.R
\name{addFreq_MitoR}
\alias{addFreq_MitoR}
\title{Updates the MitoR DB with the new patient}
\usage{
addFreq_MitoR(new_patient)
}
\arguments{
\item{patient_ID}{Patient's ID}
}
\value{
List with the new MitoR DB and the frequencies of the mutations of the new patients on the MitoR DB
}
\description{
Fixes the frequencies and count values by going through the patient's DataFrame report and the previous MitoR DB, comparing each mutation and looking for matches.
}
