% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_cnv.R
\name{analyze_cnv}
\alias{analyze_cnv}
\title{Analyze copy number variations (CNVs) of a patient sample.}
\usage{
analyze_cnv(path_dir, minoverlap = 1e-04, transit = 1)
}
\arguments{
\item{path_dir}{path of the directory of the patient that will be analyzed. It must contain R1 and R2 (fasta or fastq files).}

\item{minoverlap}{is a decimal number which indicates the needed percentage
of overlap from which it can be declare that a region has suffered a variation.}

\item{transit}{transition probability: Transition probability of the hidden Markov Chain from the normal copy number state to either a deletion or a duplication.
This parameter is taken from ExomeDepth algorithms.}
}
\value{
a list:
\itemize{
\item Patients DB (updated if the patient that has been analyzed is new).
\item CNV calls (dataframe with information about CNVs).
\item GenesDB (updated data base with information about the genes that has been detected as variant.
\item Graph with the duplications colored in green and deletions colored in red.
}
}
\description{
Generate an excel report with 2 sheets: in the first one there is
a table with information about the CNVs detected such as location, size, and hyperlinks to online data bases.
On the second sheet there is a visualization of the CNVs within the whole mitochondrial genome of the patient.
}
\examples{
outAnalisis <- analyze_cnv(path_dir = "~/MitoPatients/46608", transit = 0.7)
PatientsDB <- outAnalisis[[1]]
CNV_calls <- outAnalisis[[2]]
GenesDB <- outAnalisis[[3]]
graph <- outAnalisis[[4]]
}
