% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedSNP.R, R/bedfileMito.R
\name{bedfileMito}
\alias{bedfileMito}
\title{Load bed file}
\usage{
bedfileMito()

bedfileMito()
}
\description{
Load a dataset with information about the 37 genes within the mitochondrial genome.

Load a dataset with information about the 37 genes within the mitochondrial genome.
}
\examples{
 data("bedfileMito")
 data("bedfileMito")
}
