% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_to_bam.R
\name{fasta_to_bam}
\alias{fasta_to_bam}
\title{Generation of BAM file from FASTA or FASTQ files.}
\usage{
fasta_to_bam(path_dir)
}
\arguments{
\item{path_dir}{path of the directory of the patient that will be analyzed.}
}
\value{
path of the sorted BAM file.
}
\description{
Given the path of a directory, this function will identify the R1 and R2 files
and will generate a SAM file, then a BAM and a sorted BAM file by applying commands from BWA,
Picard, Samtools and GATK and using the mitochondrial human genome reference.
All this softwares and reference are downloaded automathically with the installation of MitoR.
}
