% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_Indel_report.R
\name{fix_filter_values}
\alias{fix_filter_values}
\title{Fix the filter parameters for GATK}
\usage{
fix_filter_values(
  QD_SNPS,
  FS_SNPS,
  MQ_SNPS,
  MQRankSum_SNPS,
  ReadPosRankSum_SNPS,
  QD_INDELS,
  FS_INDELS,
  ReadPosRankSum_INDELS
)
}
\value{
List of the same parameters it had as input
}
\description{
Modifies the parameters to the exact way GATK needs to understand them
}
