% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cnv_report.R
\name{generate_links}
\alias{generate_links}
\title{Generation of Links}
\usage{
generate_links(CNV_calls)
}
\arguments{
\item{CNV_calls}{dataframe with CNVs detected.}
}
\value{
CNV_calls with 2 new columns "Franklin" and "Varsome".
}
\description{
Generates hyperlinks that redirects the user to databases that provide more information about each variation, such as the pathogenicity and clinical implications.
}
