% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_gene (1).R
\name{generateTXT_gene}
\alias{generateTXT_gene}
\title{Check which genes have mutations.}
\usage{
generateTXT_gene(numPaciente)
}
\arguments{
\item{numPaciente}{Patient ID.}
}
\value{
Updated BED table as a DataFrame object. Containing colomns: Interval, gene name, mutations found (REF-POS-ALT)
}
\description{
Using a universal BED file and the mutations of the patient detected by MitoR, the function gives a list
of mutations, which gene has them and the interval of those genes.
}
