% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3d_analysis.R
\name{get_coords}
\alias{get_coords}
\title{Get 3D Coordinates of Protein Variants}
\usage{
get_coords(prot_list, t = NULL, target = NULL, ref = "NC_012920")
}
\arguments{
\item{prot_list}{List. Protein list generated by translating the output of \emph{extract_cds}.}

\item{t}{Double. Frequency threshold, select positions in the MSA which its most conserved amino acid/nucleotide has a frequency of at most \emph{t}.}

\item{target}{Character. Optional, if specified, the function returns the variants between the reference and the target sequence.}

\item{ref}{Character. Reference sequence.}
}
\value{
Data frame containing the 3D coordinates of the positions of non-synonymous variants. It can be
used by the function \emph{map_variants} to visualize the sites in 3D
}
\description{
Get 3D Coordinates of Protein Variants
}
