% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitor_analysis.R
\name{mitor_analysis}
\alias{mitor_analysis}
\title{Analyze mutations and variations}
\usage{
mitor_analysis(
  path_dir,
  cnv_minoverlap = 1e-04,
  cnv_transit = 1,
  QD_SNPS = "< 2.0",
  FS_SNPS = "> 60.0",
  MQ_SNPS = "< 40.0",
  MQRankSum_SNPS = "< -12.5",
  ReadPosRankSum_SNPS = "< -8.0",
  QD_INDELS = "< 2.0",
  FS_INDELS = "> 200.0",
  ReadPosRankSum_INDELS = "< -20.0"
)
}
\arguments{
\item{path_dir}{path of the directory of the patient that will be analyzed. It must contain R1 and R2 (fasta or fastq files).}

\item{minoverlap}{is a decimal number which indicates the needed percentage
of overlap from which it can be declare that a region has suffered a variation.}

\item{transit}{transition probability: Transition probability of the hidden Markov Chain
from the normal copy number state to either a deletion or a duplication.}
}
\value{
a list: Patients DB (updated if the patient that has been analyzed is new),
CNV calls (dataframe with information about CNVs),
GenesDB (updated data base with information about the genes that has been detected
as variant)
}
\description{
Generates the complete analysis to detect SNPs, indels and CNVs.
}
\examples{
outAnalisis <- MitoRAnalysis(path_dir = "~/PacientesMito/46608/BM22-46608_R1.fastq.gz")
PatientsDB<-outAnalisis[[1]]
CNV_calls<-outAnalisis[[2]]
all.exons<-outAnalisis[[3]]
}
