% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_gene.R
\name{pileupPlot}
\alias{pileupPlot}
\title{Generate a PileUp plot of the mutated genes and the mutations themselves.}
\usage{
pileupPlot(BAM_file, plot_mutation = TRUE, range = 5)
}
\arguments{
\item{BAM_file}{Patient's BAM file.}

\item{plot_mutation}{Whether to plot each mutation or not}

\item{range}{Amount of nucleotides to show next to the mutated one. plot_mutation must be TRUE in this case.}
}
\value{
Nested list with the plots. The main list is named after the mutated genes. Inside of them there is the gene plot and all mutation plots belonging to that gene.
}
\description{
Create a visual representation of the read coverage and amount of mutations per gene.
Because of the large amount of positions on each gene, it also creates a pileup plot for each mutation and a range of near nucleotides.
}
\examples{
pileupPlot("47286", plot_mutation = FALSE)
pileupPlot("47286", plot_mutation = TRUE)
pileupPlot("47286", plot_mutation = TRUE, range = 10)
}
