% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cnv.R
\name{plot_cnv}
\alias{plot_cnv}
\title{Generation of the plot for CNVs.}
\usage{
plot_cnv(CNV_calls)
}
\arguments{
\item{CNV_calls}{dataframe that contains the information about the CNVs detected.}
}
\value{
graph
}
\description{
Generates a graph that represents the CNVs detected. The user will be able to
visualize the distribution, sizes, and positions of the variations within the whole mtDNA
and the genes that are affected by them.
}
