% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitomap.R
\name{search_allele}
\alias{search_allele}
\title{Title}
\usage{
search_allele(pos = NULL)
}
\arguments{
\item{pos}{}
}
\description{
Title
}
\examples{
\dontrun{
  pos <- 13800:13900
 pos <- 5
 test <- search_allele(pos)
 test
  }
}
