% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_cnv.R
\name{update_DB_cnv}
\alias{update_DB_cnv}
\title{Add another patient's information to Genes DB}
\usage{
update_DB_cnv(CNV_calls, id, ControlDB, PatientsDB, GenesDB)
}
\arguments{
\item{CNV_calls}{dataframe which contains the information about the CNVs of the new patient.}

\item{id}{number or character that indicates which patient is.}

\item{ControlDB}{database of control patients.}

\item{PatientsDB}{database of patients to analyze or analyzed.}

\item{GenesDB}{database of genes affected.}
}
\value{
Genes DB updated.
}
\description{
Adds the information about the genes affected by the CNVs of a new patient
and keeps track of the frequency and type of variation of each gene.
}
